+++ chartlets.js/src/lib/actions/helpers/getInputValues.ts
@@ -2,16 +2,17 @@ import type { Input } from "@/lib/types/model/channel";
 import type { ContributionState } from "@/lib/types/state/contribution";
 import type { ComponentState } from "@/lib/types/state/component";
 import { isContainerState } from "@/lib/actions/helpers/isContainerState";
-import { getValue } from "@/lib/utils/objPath";
+import { getValue, toObjPath } from '@/lib/utils/objPath';
 import { isObject } from "@/lib/utils/isObject";
 
 export function getInputValues<S extends object = object>(
   inputs: Input[],
   contributionState: ContributionState,
   hostState?: S | undefined,
+  getDerivedHostState?: (hostState: object, property: string) => unknown,
 ): unknown[] {
   return inputs.map((input) =>
-    getInputValue(input, contributionState, hostState),
+    getInputValue(input, contributionState, hostState, getDerivedHostState),
   );
 }
 
@@ -20,7 +21,8 @@ const noValue = {};
 export function getInputValue<S extends object = object>(
   input: Input,
   contributionState: ContributionState,
-  hostState?: S | undefined,
+  hostState?: S,
+  getDerivedHostState?: (hostState: object, propertyName: string) => unknown,
 ): unknown {
   let inputValue: unknown = undefined;
   const dataSource = input.link || "component";
@@ -29,8 +31,7 @@ export function getInputValue<S extends object = object>(
   } else if (dataSource === "container" && contributionState.container) {
     inputValue = getInputValueFromState(input, contributionState.container);
   } else if (dataSource === "app" && hostState) {
-    console.log();
-    inputValue = getInputValueFromState(input, hostState);
+    inputValue = getInputValueFromState(input, hostState, getDerivedHostState);
   } else {
     console.warn(`input with unknown data source:`, input);
   }
@@ -64,14 +65,21 @@ export function getInputValueFromComponent(
 // we export for testing only
 export function getInputValueFromState(
   input: Input,
-  state: unknown,
-): unknown | undefined {
+  state: object | undefined,
+  getDerivedState?: (state: object, propertyName: string) => unknown,
+): unknown {
   let inputValue: unknown = state;
   if (input.id && isObject(inputValue)) {
     inputValue = inputValue[input.id];
   }
   if (isObject(inputValue)) {
-    inputValue = getValue(inputValue, input.property);
+    const state = inputValue;
+    const property = toObjPath(input.property);
+    inputValue = getValue(state, property);
+    if (inputValue === undefined && getDerivedState !== undefined) {
+      const propertyName = property.map(v => typeof v === "number" ? v.toString() : v).join(".");
+      inputValue = getDerivedState(state, propertyName);
+    }
   }
   return inputValue;
 }