+++ chartlets.js/src/lib/types/state/store.ts
@@ -12,6 +12,8 @@ import type { LoggingOptions } from "@/lib/actions/helpers/configureLogging";
 export interface FrameworkOptions<S extends object = object> {
   /** The host applications state management store. */
   hostStore?: StoreApi<S>;
+  /** Get a derived state from given host state. */
+  getDerivedHostState?: <S>(hostState: S, propertyName: string) => unknown;
   /** API options to configure backend. */
   api?: ApiOptions;
   /** Logging options. */