+++ CHANGES.md
@@ -1,3 +1,10 @@
+## Version 0.0.17 (in development)
+
+* Enhanced interface `FrameworkOptions` by property `getDerivedHostState`,
+  which is a user-supplied function that can compute derived
+  host state property. 
+  
+
 ## Version 0.0.16 (from 2024/11/12)
 
 Initial, still experimental version.